Moved another create-exe test into "create_exe.rs"
